requirements.txt
pandas>=2.2.2
pyyaml>=6.0.1
pytest>=8.2.0
boto3
python-dotenv
pyfiglet==1.0.2